[Android] Removed deps on chrome_java in c.b.search_engines
Chromium change:
https://chromium.googlesource.com/chromium/src/+/26eb7c14f9b19ae819bb7dc9a8d10e865a43fcc8

Android: Remove deps on chrome_java in c.b.search_engines

The package c.b.search_engines is used by several components in core
chrome.  This CL removes its dependencies on chrome_java to pave
the way for the other components to be modularized easily.

Main dependencies handled are:
- SettingsLauncherImpl: Injected from ToolbarManager. This is
      a stateless class so it can be shared/reused where the code
      can be simplified.
- LocaleManager: Injects a runnable wrapping
      LocaleManager#setSearchEngineAutoSwitch(false) to
      SearchEngineAdapter

Bug: 1171512
