+++ android/java/org/chromium/chrome/browser/search_engines/settings/BraveSearchEnginePreference.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.chromium.chrome.browser.settings;
+package org.chromium.chrome.browser.search_engines.settings;
 
 import android.os.Bundle;
 import android.view.View;
@@ -11,11 +11,13 @@ import android.widget.ListView;
 
 import androidx.fragment.app.ListFragment;
 
-import org.chromium.chrome.R;
-import org.chromium.chrome.browser.search_engines.settings.SearchEngineAdapter;
+import org.chromium.chrome.browser.search_engines.R;
 
-public class BraveSearchEnginePreference extends ListFragment {
+public class BraveSearchEnginePreference extends SearchEngineSettings {
+    // These members will be deleted in bytecode, member from parent class will be used instead.
     private SearchEngineAdapter mSearchEngineAdapter;
+
+    // Own members.
     private boolean mPrivate;
 
     public BraveSearchEnginePreference(boolean isPrivate) {
@@ -27,27 +29,10 @@ public class BraveSearchEnginePreference extends ListFragment {
         super.onCreate(savedInstanceState);
         getActivity().setTitle(mPrivate ? R.string.prefs_private_search_engine
                                         : R.string.prefs_standard_search_engine);
+    }
+
+    public void createAdapterIfNecessary() {
+        if (mSearchEngineAdapter != null) return;
         mSearchEngineAdapter = new BraveSearchEngineAdapter(getActivity(), mPrivate);
-        setListAdapter(mSearchEngineAdapter);
-    }
-
-    @Override
-    public void onViewCreated(View view, Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-        ListView listView = getListView();
-        listView.setDivider(null);
-        listView.setItemsCanFocus(true);
-    }
-
-    @Override
-    public void onStart() {
-        super.onStart();
-        mSearchEngineAdapter.start();
-    }
-
-    @Override
-    public void onStop() {
-        super.onStop();
-        mSearchEngineAdapter.stop();
     }
 }
