+++ android/java/org/chromium/chrome/browser/search_engines/settings/BraveStandardSearchEnginePreference.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.chromium.chrome.browser.settings;
+package org.chromium.chrome.browser.search_engines.settings;
 
 public class BraveStandardSearchEnginePreference extends BraveSearchEnginePreference {
     public BraveStandardSearchEnginePreference() {
