+++ patches/chrome-browser-search_engines-android-BUILD.gn.patch
@@ -1,12 +1,12 @@
 diff --git a/chrome/browser/search_engines/android/BUILD.gn b/chrome/browser/search_engines/android/BUILD.gn
-index ff0be3d69d686e903a58f38b4d2b41c0c7d89f4d..b56d9c1d6cdfafec88730e4d355f32653b9bf6cf 100644
+index ff0be3d69d686e903a58f38b4d2b41c0c7d89f4d..2a09a53041637f410c9fa0b7e2f8f2f896b2477f 100644
 --- a/chrome/browser/search_engines/android/BUILD.gn
 +++ b/chrome/browser/search_engines/android/BUILD.gn
-@@ -14,6 +14,7 @@ android_library("java") {
-     "java/src/org/chromium/chrome/browser/search_engines/settings/SearchEngineAdapter.java",
-     "java/src/org/chromium/chrome/browser/search_engines/settings/SearchEngineSettings.java",
+@@ -36,6 +36,7 @@ android_library("java") {
+     "//third_party/androidx:androidx_fragment_fragment_java",
+     "//ui/android:ui_full_java",
    ]
-+  sources += brave_browser_search_engines_android_sources
- 
-   srcjar_deps = [ "//components/search_engines:search_engine_type_java" ]
- 
++  sources += brave_browser_search_engines_android_sources deps += brave_browser_search_engines_android_resources_deps
+   annotation_processor_deps = [ "//base/android/jni_generator:jni_processor" ]
+   resources_package = "org.chromium.chrome.browser.search_engines"
+ }
